requirements.txt
# cd Well-Bot_SER
# .\venv\Scripts\Activate.ps1
# python test\test_ser_queue_recording.py
# python -m pip install -r requirements.txt
# py -3.12 -m venv venv
# https://well-bot-emotionrecognition-520080168829.asia-south1.run.app/ser/dashboard

# FastAPI Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Speech & Audio Processing
transformers==4.41.1
sentencepiece>=0.2.0,<1.0.0
# Note: torch and torchaudio are installed separately in Dockerfile with CPU-only version
# to save memory. They are NOT listed here to avoid syntax issues with requirements.txt.
# The Dockerfile installs: torch==2.3.0 torchaudio==2.3.0 --index-url https://download.pytorch.org/whl/cpu
librosa==0.10.1
soundfile==0.12.1
huggingface-hub==0.23.1
protobuf==3.20.3
langdetect==1.0.9
kagglehub==0.2.0

# FunASR and emotion2vec+ (Speech Emotion Recognition)
funasr>=1.0.0
modelscope>=1.0.0

# Supabase client (REST API)
supabase>=2.4.0

# Environment Configuration
pydantic==2.7.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Machine Learning & Preprocessing
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# Testing
pytest==8.2.0
httpx[http2]>=0.26,<0.28
python-multipart
